added a new test for timestamps in insertBulk, and misc code cleanup
